requirements/test.txt
# Test dependencies go here.
-r base.txt



coverage==4.1
django-coverage-plugin==1.3.1
flake8==2.6.0
django-test-plus==1.0.13
factory_boy==2.7.0

# pytest
pytest-django==2.9.1
pytest-sugar==0.7.1

# nose
nose==1.3.7
django-nose==1.4.4
